requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
importlib-metadata==0.18
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pyparsing==2.4.0
pytest==4.6.3
pytest-django==3.5.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
uuid==1.30
wcwidth==0.1.7
zipp==0.5.1
